not replace matches files

+++ system/modules/Migrations/objects/Parser/Object/Value.php
@@ -75,15 +75,21 @@ class Value extends \Migrations\Parser
         }
         switch ($type) {
             case 'image':
+                $notEq = true;
                 $dir = pathinfo($this->reader->source, PATHINFO_DIRNAME);
                 if ($this->model->{$this->param->value}) {
                     $file = \Files\File::get($this->model->{$this->param->value});
-                    if ($file) {
+                    if ($file && $value && file_exists($dir . '/' . $value) && file_exists(App::$primary->path . $file->path) && md5_file($dir . '/' . $value) == md5_file(App::$primary->path . $file->path)) {
+                        $notEq = false;
+                    }
+                    if ($file && $notEq) {
                         $file->delete();
                         $this->model->{$this->param->value} = 0;
                     }
                 }
-                $this->model->{$this->param->value} = \App::$primary->files->uploadFromUrl($dir . '/' . $value, ['accept_group' => 'image', 'upload_code' => 'MigrationUpload']);
+                if ($notEq) {
+                    $this->model->{$this->param->value} = \App::$primary->files->uploadFromUrl($dir . '/' . $value, ['accept_group' => 'image', 'upload_code' => 'MigrationUpload']);
+                }
                 break;
             default:
                 $this->model->{$this->param->value} = $value;
